Api/requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.3.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
firebase-admin==5.2.0
Flask==2.0.2
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
gensim==4.1.2
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
libcst==0.3.23
MarkupSafe==2.0.1
msgpack==1.0.2
mypy-extensions==0.4.3
nltk==3.6.5
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pluggy==1.0.0
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.8.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
rsa==4.7.2
scipy==1.7.2
six==1.16.0
smart-open==5.2.1
toml==0.10.2
tqdm==4.62.3
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.7
uvicorn==0.16.0
Werkzeug==2.0.2
gunicorn==20.1.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.1
grpcio-status==1.41.1
gevent==21.12.0